# Try to optimise OPAL