--- a/.gitignore
+++ b/.gitignore
@@ -37,3 +37,5 @@ Temporary Items
 # usermanual output
 usermanual/output
 usermanual/bin
+
+optbuild/**