--- a/sources/Adapters/picoTracker/main/CMakeLists.txt
+++ b/sources/Adapters/picoTracker/main/CMakeLists.txt
@@ -14,5 +14,6 @@ target_include_directories(picoTracker PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 include_directories(${PROJECT_SOURCE_DIR})
 
+
 # create map/bin/hex/uf2 file etc.
 pico_add_extra_outputs(picoTracker)