--- a/sources/Externals/opal/CMakeLists.txt
+++ b/sources/Externals/opal/CMakeLists.txt
@@ -2,8 +2,8 @@ add_library(opal
 opal.h opal.cpp
 )
 
-#target_link_libraries(yxml PUBLIC system_filesystem
-#)
+target_link_libraries(opal PUBLIC application_utils
+)
 
 target_include_directories(opal PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})