--- a/sources/Externals/opal/CMakeLists.txt
+++ b/sources/Externals/opal/CMakeLists.txt
@@ -2,7 +2,9 @@ add_library(opal
 opal.h opal.cpp
 )
 
-target_link_libraries(opal PUBLIC application_utils
+target_link_libraries(opal 
+    PUBLIC application_utils
+    PUBLIC hardware_flash
 )
 
 target_include_directories(opal PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})