--- a/sources/Externals/opal/opal.cpp
+++ b/sources/Externals/opal/opal.cpp
@@ -1,4 +1,5 @@
 #include "opal.h"
+#include "hardware/flash.h"
 
 const uint16_t Opal::RateTables[4][8] = {
     {1, 0, 1, 0, 1, 0, 1, 0},
@@ -7,7 +8,7 @@ const uint16_t Opal::RateTables[4][8] = {
     {1, 0, 0, 0, 0, 0, 0, 0},
 };
 //--------------------------------------------------------------------------------------------------
-const uint16_t Opal::ExpTable[0x100] = {
+static const uint16_t __not_in_flash("mydata") ExpTable[0x100] = {
     1018, 1013, 1007, 1002, 996, 991, 986, 980, 975, 969, 964, 959, 953, 948,
     942,  937,  932,  927,  921, 916, 911, 906, 900, 895, 890, 885, 880, 874,
     869,  864,  859,  854,  849, 844, 839, 834, 829, 824, 819, 814, 809, 804,
@@ -28,8 +29,9 @@ const uint16_t Opal::ExpTable[0x100] = {
     48,   45,   42,   40,   37,  34,  31,  28,  25,  22,  20,  17,  14,  11,
     8,    6,    3,    0,
 };
+
 //--------------------------------------------------------------------------------------------------
-const uint16_t Opal::LogSinTable[0x100] = {
+static const uint16_t __not_in_flash("mydata") LogSinTable[0x100] = {
     2137, 1731, 1543, 1419, 1326, 1252, 1190, 1137, 1091, 1050, 1013, 979, 949,
     920,  894,  869,  846,  825,  804,  785,  767,  749,  732,  717,  701, 687,
     672,  659,  646,  633,  621,  609,  598,  587,  576,  566,  556,  546, 536,
@@ -896,7 +898,7 @@ Opal::Operator::Output(uint16_t /*keyscalenum*/, uint32_t phase_step,
   case 0:
     if (phase & 0x100)
       offset ^= 0xFF;
-    logsin = Master->LogSinTable[offset];
+    logsin = LogSinTable[offset];
     negate = (phase & 0x200) != 0;
     break;
 
@@ -908,7 +910,7 @@ Opal::Operator::Output(uint16_t /*keyscalenum*/, uint32_t phase_step,
       offset = 0;
     else if (phase & 0x100)
       offset ^= 0xFF;
-    logsin = Master->LogSinTable[offset];
+    logsin = LogSinTable[offset];
     break;
 
   //------------------------------------
@@ -917,7 +919,7 @@ Opal::Operator::Output(uint16_t /*keyscalenum*/, uint32_t phase_step,
   case 2:
     if (phase & 0x100)
       offset ^= 0xFF;
-    logsin = Master->LogSinTable[offset];
+    logsin = LogSinTable[offset];
     break;
 
   //------------------------------------
@@ -926,7 +928,7 @@ Opal::Operator::Output(uint16_t /*keyscalenum*/, uint32_t phase_step,
   case 3:
     if (phase & 0x100)
       offset = 0;
-    logsin = Master->LogSinTable[offset];
+    logsin = LogSinTable[offset];
     break;
 
   //------------------------------------
@@ -945,7 +947,7 @@ Opal::Operator::Output(uint16_t /*keyscalenum*/, uint32_t phase_step,
       negate = (phase & 0x100) != 0;
     }
 
-    logsin = Master->LogSinTable[offset];
+    logsin = LogSinTable[offset];
     break;
 
   //------------------------------------
@@ -962,7 +964,7 @@ Opal::Operator::Output(uint16_t /*keyscalenum*/, uint32_t phase_step,
         offset ^= 0xFF;
     }
 
-    logsin = Master->LogSinTable[offset];
+    logsin = LogSinTable[offset];
     break;
 
   //------------------------------------
@@ -996,7 +998,7 @@ Opal::Operator::Output(uint16_t /*keyscalenum*/, uint32_t phase_step,
   // (the hidden bit) is then the significand of the floating point output and
   // the yet unused MSB's of the input are the exponent of the floating point
   // output."
-  int16_t v = (Master->ExpTable[mix & 0xFF] + 1024u) >> (mix >> 8u);
+  int16_t v = (ExpTable[mix & 0xFF] + 1024u) >> (mix >> 8u);
   v += v;
   if (negate)
     v = ~v;