--- a/sources/Externals/opal/opal.cpp
+++ b/sources/Externals/opal/opal.cpp
@@ -7,8 +7,10 @@ const uint16_t Opal::RateTables[4][8] = {
     {1, 0, 0, 0, 1, 0, 0, 0},
     {1, 0, 0, 0, 0, 0, 0, 0},
 };
+// These look up tables are accessed in tight loops for each sample, so they
+// need to be in RAM not flash for fast access
 //--------------------------------------------------------------------------------------------------
-static const uint16_t __not_in_flash("mydata") ExpTable[0x100] = {
+static const uint16_t __not_in_flash("opalLUTdata") ExpTable[0x100] = {
     1018, 1013, 1007, 1002, 996, 991, 986, 980, 975, 969, 964, 959, 953, 948,
     942,  937,  932,  927,  921, 916, 911, 906, 900, 895, 890, 885, 880, 874,
     869,  864,  859,  854,  849, 844, 839, 834, 829, 824, 819, 814, 809, 804,
@@ -31,7 +33,7 @@ static const uint16_t __not_in_flash("mydata") ExpTable[0x100] = {
 };
 
 //--------------------------------------------------------------------------------------------------
-static const uint16_t __not_in_flash("mydata") LogSinTable[0x100] = {
+static const uint16_t __not_in_flash("opalLUTdata") LogSinTable[0x100] = {
     2137, 1731, 1543, 1419, 1326, 1252, 1190, 1137, 1091, 1050, 1013, 979, 949,
     920,  894,  869,  846,  825,  804,  785,  767,  749,  732,  717,  701, 687,
     672,  659,  646,  633,  621,  609,  598,  587,  576,  566,  556,  546, 536,