--- a/sources/Externals/opal/opal.h
+++ b/sources/Externals/opal/opal.h
@@ -24,6 +24,7 @@
 #ifndef _OPAL_H_
 #define _OPAL_H_
 
+#include "Application/Utils/fixed.h"
 #include <cstdint>
 
 //==================================================================================================
@@ -184,6 +185,8 @@ public:
   void SetSampleRate(int sample_rate);
   void Port(uint16_t reg_num, uint8_t val);
   void Sample(int16_t *left, int16_t *right);
+
+  void SampleBuffer(fixed *buffer, int size);
 
 protected:
   void Init(int sample_rate);