--- a/sources/Externals/opal/opal.h
+++ b/sources/Externals/opal/opal.h
@@ -197,8 +197,6 @@ protected:
   int16_t LastOutput[2], CurrOutput[2];
   Channel Chan[NumChannels];
   Operator Op[NumOperators];
-  //      uint16_t            ExpTable[256];
-  //      uint16_t            LogSinTable[256];
   uint16_t Clock;
   uint16_t TremoloClock;
   uint16_t TremoloLevel;
@@ -209,8 +207,6 @@ protected:
   bool VibratoDepth;
 
   static const uint16_t RateTables[4][8];
-  static const uint16_t ExpTable[256];
-  static const uint16_t LogSinTable[256];
 };
 
 #endif