As part of trying to improve performance for #283 this PR does some basic optimisation for the OPAL audio rendering code. The main optimisation in the OPAL code is removing function calls from the tight loop by creating a new method for generating samples together for a whole buffer and then further removing more function calls by inlining several small functions. At the cost of 200bytes of ram, the constant LUTs are also forced to be in ram all the time.

For a release builds takes rendering time for 1 `Render()` method call at default bpm and default OPAL instrument settings from:

```
Render took: 3838 (0 ts)[268860620]
Render took: 3820 (0 ts)[268860620]
Render took: 3813 (0 ts)[268860620]
Render took: 3764 (0 ts)[268860620]
Render took: 3800 (0 ts)[268860620]
Render took: 3811 (0 ts)[268860620]
Render took: 3822 (0 ts)[268860620]
Render took: 3802 (0 ts)[268860620]
Render took: 3773 (0 ts)[268860620]
Render took: 3827 (0 ts)[268860620]
```

to:
```
[OPALINSTRUMENT] Render took: 2832 us [799])
[OPALINSTRUMENT] Render took: 2812 us [799])
[OPALINSTRUMENT] Render took: 2783 us [799])
[OPALINSTRUMENT] Render took: 2829 us [799])
[OPALINSTRUMENT] Render took: 2827 us [799])
[OPALINSTRUMENT] Render took: 2805 us [798])
[OPALINSTRUMENT] Render took: 2786 us [799])
[OPALINSTRUMENT] Render took: 2792 us [799])
[OPALINSTRUMENT] Render took: 2813 us [799])
[OPALINSTRUMENT] Render took: 2840 us [799])
[OPALINSTRUMENT] Render took: 2814 us [799])
```

so from around avg of `3.8ms` to `2.8ms` which is a decent if not stellar improvement.