@maks 2025-01-02
> Very nice!
> 
> Wouldn't inlining `Sample` and `Output` methods in the current code achieve the same results?

Unfortunately no as the linker complains if you try to call an inline function, in the this case `Sample()` from another compilation object, which sort of makes sense because the compiler has removed it as its already been inlined (ie. deleted) from its own compilation object.

> Do you have a sense on how much of this gain is due to the method change vs putting that data in memory? (BTW, wouldn't you achieve not being in flash by not defining it `const`?)

Good question! I just reran test with the force into ram macro removed from the LUTs and I get:
```
[OPALINSTRUMENT] Render took: 3306 us [799])
[OPALINSTRUMENT] Render took: 3292 us [799])
[OPALINSTRUMENT] Render took: 3249 us [799])
[OPALINSTRUMENT] Render took: 3310 us [799])
[OPALINSTRUMENT] Render took: 3337 us [799])
[OPALINSTRUMENT] Render took: 3304 us [799])
[OPALINSTRUMENT] Render took: 3280 us [799])
```

so moving the LUT data into ram is what gives approx half the performance improvement.