@maks 2025-01-03
> Wouldn't inlining `Sample` and `Output` methods in the current code achieve the same results?

Oops! sorry I see what you mean now, make them inline and then call them from the new SampleBuffer function instead.
Yes that works and just tested gives same perf improvement 👍🏻